Multiple regression on personality types and belief in conspiracy theories with python
This was done for my thesis in my final year of applied psychology. I wanted to explore if belief in conspiracy theories would be correlated with peoples personality types (They're not really!).  I used a Microsoft form which was shared on my instagram to gather the data. The form was essentially a short Big Five personality type and a conspiracy belief Likert scale. I originally used a bit of Excel and SPSS to do the analysis, but decided I should redo it to show off some data cleaning and statics skills with python. If you are interested in the thesis and would like to read a bit more, feel free to get in touch and I'll send on the full Word Doc!