An independent T-test on score data
In this project, I was given data on participants' scores at the game 'Tetris'. They were distracted with either an audio or visual interruption. My task was to examine if there is a significant difference in participants score, depending on their interruption type. The project highlights some data cleaning and descriptive and inferential statistic skills!